+++ src/lib.rs
@@ -10,6 +10,9 @@ pub use sms::SmsData;
 
 #[derive(Debug)]
 pub enum AmlError {
+    /// You have tried to parse an unimplemented version of SMS AML.
     UnimplementedVersion,
+
+    /// You have tried to parse an corrumpted base64 SMS data.
     InvalidBase64,
 }