+++ CHANGELOG.md
@@ -1,8 +1,17 @@
+# Changelog
+
+## 0.1.2 - 2021-06-04
+
+### Added
+
+- `is_validated` property add to SmsData implementation.
+- Improvement of documentation.
+
 ## 0.1.1 - 2021-06-03
 
 ### Added
 
-- Authentification via HMAC-SHA1 added to HTTPS message. 
+- Authentification via HMAC-SHA1 added to HTTPS message.
 
 ## 0.1.0 - 2021-05-29